Added doc strings.

+++ lib/vgrepo/repository.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# encoding: utf8
+# coding: utf8
 
 import hashlib
 import json
@@ -12,6 +12,12 @@ from packaging.version import Version
 from .meta.images import VMetadataImage
 
 
+class VImageNotFound(Exception):
+
+    def __init__(self, path):
+        self.path = path
+
+
 class VImageVersionFoundError(Exception):
 
     def __init__(self, name, version):
@@ -20,21 +26,40 @@ class VImageVersionFoundError(Exception):
 
 
 class VRepository:
+    """
+    Provides methods to manage repository in the storage
+    """
 
     SHA1_BUFFER_SIZE = 65536
 
     @property
     def is_empty(self):
+        """
+        Returns is the repository empty (does not contain any versions) or not
+        
+        :return: bool
+        """
         return self.meta.versions is None or self.meta.versions == []
 
     @property
     def has_meta(self):
+        """
+        Returns is the repository has metadata saved on the disk or not
+        
+        :return: bool
+        """
         try:
             return os.path.isfile(self.meta_path)
         except (OSError, IOError):
-            print("unable to read metadata for '{0}'".format(self.meta.name))
+            print("Error: unable to read metadata for '{0}'".format(self.meta.name))
 
     def has_image(self, version=None):
+        """
+        Returns is the image present or not by given version
+        
+        :param version: version of the image
+        :return: bool
+        """
         try:
             if version:
                 path = self.get_image_path(version)
@@ -45,41 +70,89 @@ class VRepository:
         except (OSError, IOError):
             print("Error: unable to read image '{0}'".format(self.meta.name))
 
+        return False
+
     def is_exist(self, version=None):
+        """
+        Returns is the repository:
+        a) Has their own metadata on the disk
+        b) Has saved image file in the repository
+        
+        :param version: version of the image 
+        :return: bool
+        """
         return self.has_meta and self.has_image(version)
 
     @property
     def meta_dir(self):
+        """
+        Returns metadata directory for the repository
+        
+        :return: str
+        """
         return os.path.join(self.image_dir, "metadata")
 
     @property
     def meta_path(self):
+        """
+        Returns full metadata path for the repository 
+        
+        :return: 
+        """
         path_format = "{name}.json"
 
         return os.path.join(self.meta_dir, path_format.format(name=self.meta.name))
 
     @property
     def image_dir(self):
+        """
+        Returns directory for the images of the repository
+        
+        :return: str
+        """
         return os.path.join(self.settings.storage_path, self.meta.name)
 
     @property
     def repo_url(self):
+        """
+        Returns URL to the metadata file
+        
+        :return: 
+        """
         url_format = "{url}/{name}"
 
         return url_format.format(url=self.settings.storage_url, name=self.meta.name)
 
     def get_image_path(self, version):
+        """
+        Returns full path to the image
+        
+        :param version: version of the image
+        :return: str
+        """
         path_format = "{name}-{version}.box"
 
         return os.path.join(self.image_dir, path_format.format(name=self.meta.name, version=version))
 
     def get_image_url(self, version):
+        """
+        Returns direct URL to the image 
+        
+        :param version: version of the image
+        :return: str
+        """
         url_format = "{url}/{name}-{version}.box"
 
         return url_format.format(url=self.repo_url, name=self.meta.name, version=version)
 
     @staticmethod
     def get_sha1_checksum(path):
+        """
+        Returns SHA1 string of the file by given path
+        
+        :param path: path to the hashed file
+        :return: 
+        """
         sha1 = hashlib.sha1()
 
         try:
@@ -95,12 +168,22 @@ class VRepository:
         return sha1.hexdigest() if sha1 else sha1
 
     def load_meta(self):
+        """
+        Loads or creates metadata for itself by given name
+        
+        :return: 
+        """
         if self.has_meta:
             return VRepository.parse_meta(self.meta_path, VMetadataImage)
         else:
             return VMetadataImage(name=self.meta.name)
 
     def dump_meta(self):
+        """
+        Saves metadata on the disk
+        
+        :return: 
+        """
         path = self.meta_dir
         try:
             if not os.path.isdir(path):
@@ -116,6 +199,14 @@ class VRepository:
 
     @staticmethod
     def parse_meta(cnf, cls, is_list=False):
+        """
+        Parse metadata on JSON format to the object 
+        
+        :param cnf: path to JSON file
+        :param cls: name of class to parse
+        :param is_list: is the object list or not
+        :return: instance of the class (cls)
+        """
         with open(cnf, 'r') as stream:
             target = json.load(stream)
             if is_list:
@@ -124,14 +215,35 @@ class VRepository:
                 return cls.from_json(target)
 
     @staticmethod
-    def is_equal_versions(cur, new):
-        return Version(cur) == Version(new)
+    def is_equal_versions(first, second):
+        """
+        Returns are versions equal or not
+        
+        :param first: first version
+        :param second: second version
+        :return: bool
+        """
+        return Version(str(first)) == Version(str(second))
 
     @staticmethod
-    def not_equal_versions(cur, new):
-        return not VRepository.is_equal_versions(cur, new)
+    def not_equal_versions(first, second):
+        """
+        Returns are version not equal or not 
+        
+        :param first: first version
+        :param second: second version
+        :return: bool 
+        """
+        return not VRepository.is_equal_versions(first, second)
 
     def filter_versions(self, func, version=None):
+        """
+        Returns list of images without versions by given function filter
+        
+        :param func: name of the filter function
+        :param version: number of version
+        :return: 
+        """
         meta = deepcopy(self.meta)
 
         if not self.is_empty:
@@ -143,6 +255,11 @@ class VRepository:
         return meta
 
     def remove_meta(self):
+        """
+        Removes metadata from disk
+        
+        :return: 
+        """
         try:
             if self.has_meta:
                 os.remove(self.meta_path)
@@ -153,9 +270,21 @@ class VRepository:
         return True
 
     def sync_meta(self, meta):
+        """
+        Saves meta in memory by given copy
+        
+        :param meta: copied meta object
+        :return: 
+        """
         self.meta = deepcopy(meta)
 
     def has_version(self, version):
+        """
+        Returns is the repository has version of the image
+        
+        :param version: version of the image
+        :return: 
+        """
         entries = self.filter_versions(VRepository.is_equal_versions, version)
 
         if entries.versions:
@@ -164,9 +293,19 @@ class VRepository:
             return False
 
     def copy_image(self, src, version):
+        """
+        Copies image to the repository's directory
+        
+        :param src: path to the original image file
+        :param version: version of the image
+        :return: 
+        """
         try:
             if not os.path.isdir(self.image_dir):
                 os.makedirs(self.image_dir)
+            if not os.path.isfile(src):
+                raise VImageNotFound(src)
+
             shutil.copy2(src, self.get_image_path(version))
             return True
         except (OSError, IOError):
@@ -175,9 +314,14 @@ class VRepository:
         return False
 
     def remove_image(self, version):
+        """
+        Removes image from the repository
+        
+        :param version: version of the image
+        :return: 
+        """
         path = self.get_image_path(version)
 
-        print path
         try:
             if self.is_exist(version):
                 os.remove(path)
@@ -187,7 +331,12 @@ class VRepository:
 
         return False
 
-    def destroy_image(self):
+    def destroy(self):
+        """
+        Destroys all images from the repository
+        
+        :return: 
+        """
         try:
             shutil.rmtree(self.image_dir, ignore_errors=True)
         except (OSError, IOError):
@@ -203,6 +352,13 @@ class VRepository:
         self.meta = self.load_meta()
 
     def add(self, src, img):
+        """
+        Adds image to the repository by given file and metadata
+        
+        :param src: source image
+        :param img: image's metadata
+        :return: 
+        """
         meta, image = deepcopy(self.meta), deepcopy(img)
 
         if self.is_empty:
@@ -211,6 +367,7 @@ class VRepository:
         for v in image.versions:
             if not self.has_version(v.version):
                 for p in v.providers:
+                    p.name = p.name or "virtualbox"
                     p.checksum_type = "sha1"
                     p.checksum = VRepository.get_sha1_checksum(src)
                     p.url = self.get_image_url(v.version)
@@ -226,9 +383,20 @@ class VRepository:
 
     @property
     def info(self):
+        """
+        Returns metadata of the repository
+        
+        :return: 
+        """
         return self.meta
 
     def remove(self, version):
+        """
+        Removes image by given version
+        
+        :param version: version of the image
+        :return: 
+        """
         meta = self.filter_versions(VRepository.not_equal_versions, version)
 
         self.remove_image(version)
@@ -236,7 +404,7 @@ class VRepository:
 
         if self.is_empty:
             self.remove_meta()
-            self.destroy_image()
+            self.destroy()
             self.meta = VMetadataImage(self.meta.name)
         else:
             self.dump_meta()
